completion/requirements.txt
six
tqdm==4.30.*
future
configargparse
seutil
torchtext==0.4.0
